feat-users: implementação do caso de uso para deletar endereço de usuário, incluindo validações aprimoradas e ajustes nas mensagens de erro

+++ users/src/modules/user2/application/use-cases/delete-user.usecase.ts
@@ -9,8 +9,8 @@ export class DeleteUserUseCase {
   async execute(id: string): Promise<void> {
     const user = await this.userRepository.findOne({ userId: id });
 
-    if (!user) {
-      throw new NotFoundItem();
+    if (user == undefined || user == null) {
+      throw new NotFoundItem('Não foi possivel encontrar o usuário');
     }
 
     await this.userRepository.delete(id);
